Show mock login button only on login page.
