+++ frontend/src/styles/Layout.css
@@ -56,7 +56,7 @@
 
 .MockLoginButton {
     --icon-height: 36px;
-    bottom: 4em;
+    bottom: 0;
     left: 0;
 }
